Crate nbez

A crate that provides generic bezier curves of varying order and dimensionality, including numerous stack-allocated curves and an (admittedly not entirely functional) n-order curve.

Structs

Bez1o
Order 1 bezier curve

Bez2o
Order 2 bezier curve

Bez3o
Order 3 bezier curve

Bez4o
Order 4 bezier curve

Bez5o
Order 5 bezier curve

Bez6o
Order 6 bezier curve

BezChain
A chain of bezier curves, with the last point of each curve being the first point of the next.

BezIter
Iterator over bezier curve chains

InterpIter
An iterator over a bezier curve's interpolated points

NBez
An n-order bezier curve. The from_slice, split, and split_unbounded functions currently do not work.

Point2d
2-dimensional point

Point3d
3-dimensional point

Point4d
4-dimensional point

Vector2d
2-dimensional vector

Vector3d
3-dimensional vector

Vector4d
4-dimensional vector

Traits

BezCurve
Bezier curve trait

Float
A helper trait to simplify float generics

OrderStatic
Trait to mark curves that have order known at compiletime.

PVOps
A trait that specifies the necessary operators needed to have a point which nbez can properly perform operations on

Point
Specifies the needed traits to have a nbez point, as well as the vector type that this corresponds to

Vector
A vector. Gets associated with any number of points