Struct nbez::BezChain

pub struct BezChain<F, B, C> where F: Float, B: BezCurve<F> + OrderStatic, C: AsRef<[B::Point]> {
    // some fields omitted
}

A chain of bezier curves, with the last point of each curve being the first point of the next.

Methods

impl<F, B, C> BezChain<F, B, C> where F: Float, B: BezCurve<F> + OrderStatic, C: AsRef<[B::Point]>

fn from_container(container: C) -> BezChain<F, B, C>

Create a new BezChain by wrapping around a container.

fn get(&self, index: usize) -> Option<B>

Get the bezier curve that is index curves away from the start. Returns None if not enough points exist for the given curve index.

fn iter(&self) -> BezIter<F, B>

Get an iterator over all curves in the chain.

fn order(&self) -> usize

Get the order of the chain's curves. Identical to order_static().

fn unwrap(self) -> C

Unwrap the chain, returning the underlying container.