Struct nbez::Point3d

pub struct Point3d<F: Float> {
    pub x: F,
    pub y: F,
    pub z: F,
}

3-dimensional point

Fields
x: F y: F z: F
Methods

impl<F: Float> Point3d<F>

fn new(x: F, y: F, z: F) -> Point3d<F>

Convenience method for creating a new instance of this struct. Fields of this function correspond to fields of the struct.