Struct nbez::Point4d

pub struct Point4d<F: Float> {
    pub x: F,
    pub y: F,
    pub z: F,
    pub w: F,
}

4-dimensional point

Fields
x: F y: F z: F w: F
Methods

impl<F: Float> Point4d<F>

fn new(x: F, y: F, z: F, w: F) -> Point4d<F>

Convenience method for creating a new instance of this struct. Fields of this function correspond to fields of the struct.